Add files via upload
The goal is to generate a possible prime number that can be further tested using Miller-Rabin Primality test. This code attempts to accomplish the following: 

1. Pick a random integer smaller than 10,000 which is relatively prime to 2 · 3 · 5 · 7 · 11 =
2310. You should do this by generating a random number, n, between 1 and 10,000
and then checking gcd(n, 2310) = 1. If not, you should pick another random number
between 1 and 10,000 and repeat the procedure until you find one.

2. Pick a random integer K such that x = 2310K +n is 100-bit long (i.e. 2100 < x ≤ 2
101).
Note that 2100 ∼ 1.26 · 1030 so x is about 30-digit long number. Think about the range
of K carefully so that x is indeed 100-bit long.